use decimal everywhere

+++ src/DebtAmortizationItem.cs
@@ -2,12 +2,12 @@
 {
     public class DebtAmortizationItem
     {
-        private double CurrentBalance { get; }
+        private decimal CurrentBalance { get; }
         private decimal MonthlyRate { get; }
         public string Name => debtInfo.Name;
-        public double Payment { get; }
-        public double Interest => DebtInfo.RoundUp(MonthlyRate * (decimal)CurrentBalance, 2);
-        public double AppliedPayment => Payment - Interest;
+        public decimal Payment { get; }
+        public decimal Interest => DebtInfo.RoundUp(MonthlyRate * CurrentBalance, 2);
+        public decimal AppliedPayment => Payment - Interest;
         public decimal RemainingBalance => (decimal)(CurrentBalance - AppliedPayment);
 
         public readonly DebtInfo debtInfo;
@@ -18,10 +18,10 @@
             Payment = debt.CurrentPayment;
             CurrentBalance = debt.Balance;
             MonthlyRate = Payment >= CurrentBalance ? 0 : debt.AverageMonthyPr;
-            debtInfo = new DebtInfo(Name, (double)RemainingBalance, debt.Rate, Payment);
+            debtInfo = new DebtInfo(Name, RemainingBalance, debt.Rate, Payment);
         }
 
-        public DebtAmortizationItem(string name, double payment, double rate, double currentBalance) : this(
+        public DebtAmortizationItem(string name, decimal payment, decimal rate, decimal currentBalance) : this(
             new DebtInfo(name, currentBalance, rate, payment)) { }
 
         /// <inheritdoc />
